int type to long long type

+++ BitOperate.cpp
@@ -8,7 +8,7 @@ using namespace std;
 
 long long BitOperate::BitReserve(long long DataToReverse, long long BitLength){
     long long result = 0;
-    for(int i = 0; i < BitLength; i++){
+    for(long long i = 0; i < BitLength; i++){
         if((DataToReverse >> i) & 1){
             result |= 1 << (BitLength - 1 -i);
         }
@@ -17,9 +17,9 @@ long long BitOperate::BitReserve(long long DataToReverse, long long BitLength){
 }
 
 /*
-    int n = 8;
+    long long n = 8;
     BitOperate rev;
-    for (int i = 0; i < n; i++){
+    for (long long i = 0; i < n; i++){
         ans = rev.BitReserve(i, log2(n));
         cout << "i = " << i << " reverse index = " << ans << endl;
     }
