+++ NWC_math.h
@@ -5,13 +5,13 @@
 #include <vector>
 using namespace std;
 
-int AddMod(int, int, int);
-int SubMod(int in_1, int in_2, int q);
-int XGCD(int a, int q);
-int InvMod(int a, int q);
-int DivMod(int a,int b,int q);
-int MulMod(int a,int b,int q);
-int ExpMod(int a, int exp, int q);
+long long AddMod(long long, long long, long long);
+long long SubMod(long long in_1, long long in_2, long long q);
+long long XGCD(long long a, long long q);
+long long InvMod(long long a, long long q);
+long long DivMod(long long a,long long b,long long q);
+long long MulMod(long long a,long long b,long long q);
+long long ExpMod(long long a, long long exp, long long q);
 long long find_phi(long long n, long long modular); //// p = 1 mod 2n
 bool isPrime(long long n);
 long long find_prou(long long n, long long modular);
